// src/Services/Models/Login/SignInModel.ts
export interface SignInModel  {
    email: string;
    password: string;
  }
export interface OTPModel  {
    otp: string;
    email: string;
  }
  export interface EmailOTPModel  {
    email: string;
  }
  export interface UserModel  {
    email: string;
    first_name: string;
    last_name: string;
    middle_name: string;
    birth_date: string;
    contact_no: string;
    city: string;
    baranggay: string;
    street: string;
    bussiness_id: string;
    name: string;
    user_id: string;
    authorize: string;

  }